+++ src/ping32.h
@@ -4,12 +4,13 @@
 #define PING32_H
 
 #include <Arduino.h>
+#include <stdint.h>
 
 typedef void(*ping_recv_function)(void* arg, void *pdata);
 typedef void(*ping_sent_function)(void* arg, void *pdata);
 
 struct ping_option {
-    uint32_t count;
+    int16_t count;
     uint32_t ip;
     uint32_t coarse_time;
     ping_recv_function recv_function;
@@ -29,8 +30,8 @@ struct ping_resp {
 };
 
 bool ping_start(struct ping_option *ping_opt);
-void ping(const char *name, int count, int interval, int size, int timeout);
-bool ping_start(IPAddress adr, int count, int interval, int size, int timeout, struct ping_option *ping_o = NULL);
+void ping(const char *name, int16_t count, int interval, int size, int timeout);
+bool ping_start(IPAddress adr, int16_t count, int interval, int size, int timeout, struct ping_option *ping_o = NULL);
 
 #endif
 #endif // PING_H
